refactor(resources/about): unified folder
